format

+++ rabbit-job-admin/src/main/java/com/snail/job/admin/service/trigger/TriggerPoolService.java
@@ -22,12 +22,12 @@ public class TriggerPoolService {
      * key: 秒
      * val: 任务ID集合
      */
-    private final Map<Integer, Set<Long>> triggerJobMap = new ConcurrentHashMap<>();
+    private final Map<Long, Set<Long>> triggerJobMap = new ConcurrentHashMap<>();
 
     /**
      * 添加调度到缓存中
      */
-    public void push(Long jobId, Integer invokeSecond) {
+    public void push(Long jobId, Long invokeSecond) {
         Set<Long> jobIdSet = triggerJobMap.computeIfAbsent(invokeSecond, k -> new HashSet<>());
         jobIdSet.add(jobId);
     }
@@ -35,7 +35,7 @@ public class TriggerPoolService {
     /**
      * 从缓存中移除
      */
-    public Set<Long> get(Integer second) {
+    public Set<Long> get(Long second) {
         return triggerJobMap.remove(second);
     }
